Modify build script for GH

+++ build.sh
@@ -1,22 +1,42 @@
 #!/bin/bash
 OUTPUT_BIN="jcsystool"
 
-LINUX_BIN="$OUTPUT_BIN-linux-amd64"
-MAC_BIN="$OUTPUT_BIN-darwin-amd64"
 
-LINUX_OUT="./build/$LINUX_BIN.tar.gz"
-MAC_OUT="./build/$MAC_BIN.tar.gz"
+function release() {
+  VERSION="$1"
 
-mkdir -p ./build
+  if [ -z "$VERSION" ]; then
+    echo "Version number is required to release"
+    exit 1
+  else
+    git tag -a $VERSION -m "Release version $VERSION"
 
-env GOOS=linux GOARCH=amd64  go build -v -o $LINUX_BIN *.go
-tar -cvf $LINUX_OUT $LINUX_BIN
-rm $LINUX_BIN
+    LINUX_BIN="$OUTPUT_BIN-linux-amd64-$VERSION"
+    MAC_BIN="$OUTPUT_BIN-darwin-amd64-$VERSION"
+
+    LINUX_OUT="./build/$LINUX_BIN.tar.gz"
+    MAC_OUT="./build/$MAC_BIN.tar.gz"
+
+    mkdir -p ./build
+
+    env GOOS=linux GOARCH=amd64  go build -o $LINUX_BIN *.go
+    tar -cvf $LINUX_OUT $LINUX_BIN
+    rm $LINUX_BIN
 
 
-env GOOS=darwin GOARCH=amd64 go build -v -o $MAC_BIN *.go
-tar -cvf $MAC_OUT $MAC_BIN
-rm $MAC_BIN
+    env GOOS=darwin GOARCH=amd64 go build -o $MAC_BIN *.go
+    tar -cvf $MAC_OUT $MAC_BIN
+    rm $MAC_BIN
 
-echo "Built Linux: $LINUX_OUT"
-echo "Built Mac: $MAC_OUT"
+    LINUX_SHA=$(shasum -a 256 $LINUX_OUT)
+    MAC_SHA=$(shasum -a 256 $MAC_OUT)
+
+    echo "Built Linux: $LINUX_OUT"
+    echo $LINUX_SHA
+    echo && echo
+    echo "Built Mac: $MAC_OUT"
+    echo $MAC_SHA
+  fi
+}
+
+release $1
